refactor: split into more files
AbstractPlayerCharacter and AbstractMagicWielder are moved into their own files, and out of IPlayerCharacter.kt
