+++ src/main/kotlin/cl/uchile/dcc/finalreality/model/character/player/AbstractMagicWielder.kt
@@ -7,44 +7,39 @@
  */
 package cl.uchile.dcc.finalreality.model.character.player
 
+import cl.uchile.dcc.finalreality.exceptions.Require
 import cl.uchile.dcc.finalreality.model.character.IGameCharacter
-import java.util.*
 import java.util.concurrent.BlockingQueue
 
 /**
- * A White Mage is a type of [IPlayerCharacter] that can cast white magic.
+ * An abstract class that holds the information for a character that wields magic (and thus has Mana)
  *
  * @param name        the character's name
  * @param maxHp       the character's maximum health points
  * @param maxMp       the character's maximum magic points
  * @param defense     the character's defense
  * @param turnsQueue  the queue with the characters waiting for their turn
- * @constructor Creates a new Black Mage.
+ * @constructor Creates a new magic wielder.
  *
  * @property currentMp The current MP of the character.
  * @property currentHp The current HP of the character.
  *
- * @author <a href="https://www.github.com/r8vnhill">R8V</a>
  * @author <a href="https://www.github.com/johnny-godoy">Johnny</a>
  */
-class WhiteMage(
-  name: String,
-  maxHp: Int,
-  maxMp: Int,
-  defense: Int,
-  turnsQueue: BlockingQueue<IGameCharacter>
-) : AbstractMagicWielder(name, maxHp, maxMp, defense, turnsQueue) {
-
-    override fun equals(other: Any?) = when {
-        this === other                 -> true
-        other !is WhiteMage            -> false
-        hashCode() != other.hashCode() -> false
-        name != other.name             -> false
-        maxHp != other.maxHp           -> false
-        maxMp != other.maxMp           -> false
-        defense != other.defense       -> false
-        else                           -> true
+abstract class AbstractMagicWielder(
+    name: String,
+    maxHp: Int,
+    maxMp: Int,
+    defense: Int,
+    turnsQueue: BlockingQueue<IGameCharacter>
+) : AbstractPlayerCharacter(name, maxHp, defense, turnsQueue) {
+    val maxMp = Require.Stat(maxMp, "Max MP") atLeast 0
+    var currentMp: Int = maxMp
+        set(value) {
+            field = Require.Stat(value, "Current MP") inRange 0..maxMp
+        }
+    override fun toString(): String {
+        val superString = super.toString().dropLast(1)
+        return "$superString, maxMp=$maxMp, currentMp=$currentMp)"
     }
-
-    override fun hashCode() = Objects.hash(WhiteMage::class, name, maxHp, maxMp, defense)
 }
