+++ src/main/kotlin/cl/uchile/dcc/finalreality/model/character/player/IPlayerCharacter.kt
@@ -29,65 +29,3 @@ interface IPlayerCharacter {
    */
   fun equip(weapon: Weapon)
 }
-
-/**
- * A class that holds all the information of a player-controlled character in the game.
- *
- * @param name        the character's name
- * @param maxHp       the character's maximum health points
- * @param defense     the character's defense
- * @param turnsQueue  the queue with the characters waiting for their turn
- * @constructor Creates a new playable character.
- *
- * @author <a href="https://www.github.com/r8vnhill">R8V</a>
- * @author <a href="https://www.github.com/johnny-godoy">Johnny</a>
- */
-abstract class AbstractPlayerCharacter(
-  name: String,
-  maxHp: Int,
-  defense: Int,
-  turnsQueue: BlockingQueue<IGameCharacter>
-) : AbstractCharacter(name, maxHp, defense, turnsQueue), IPlayerCharacter {
-
-  private lateinit var _equippedWeapon: Weapon
-  override val equippedWeapon: Weapon
-    get() = _equippedWeapon
-
-  override fun equip(weapon: Weapon) {
-    _equippedWeapon = weapon
-  }
-}
-
-
-/**
- * An abstract class that holds the information for a character that wields magic (and thus has Mana)
- *
- * @param name        the character's name
- * @param maxHp       the character's maximum health points
- * @param maxMp       the character's maximum magic points
- * @param defense     the character's defense
- * @param turnsQueue  the queue with the characters waiting for their turn
- * @constructor Creates a new magic wielder.
- *
- * @property currentMp The current MP of the character.
- * @property currentHp The current HP of the character.
- *
- * @author <a href="https://www.github.com/johnny-godoy">Johnny</a>
- */
-abstract class AbstractMagicWielder(
-  name: String,
-  maxHp: Int,
-  maxMp: Int,
-  defense: Int,
-  turnsQueue: BlockingQueue<IGameCharacter>
-) : AbstractPlayerCharacter(name, maxHp, defense, turnsQueue) {
-  val maxMp = Require.Stat(maxMp, "Max MP") atLeast 0
-  var currentMp: Int = maxMp
-    set(value) {
-      field = Require.Stat(value, "Current MP") inRange 0..maxMp
-    }
-  override fun toString(): String {
-    val superString = super.toString().dropLast(1)
-    return "$superString, maxMp=$maxMp, currentMp=$currentMp)"
-  }
-}
